+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/TestInfoResources.scala
@@ -4,7 +4,7 @@ import akka.http.scaladsl.model.{HttpResponse, StatusCodes}
 import akka.http.scaladsl.server._
 import akka.util.Timeout
 import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport
-import pl.touk.nussknacker.engine.definition.TestingCapabilities
+import pl.touk.nussknacker.engine.definition.test.TestingCapabilities
 import pl.touk.nussknacker.restmodel.displayedgraph.DisplayableProcess
 import pl.touk.nussknacker.ui.process.repository.FetchingProcessRepository
 import pl.touk.nussknacker.ui.process.test.ScenarioTestService