+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/ManagementResourcesSpec.scala
@@ -283,13 +283,12 @@ class ManagementResourcesSpec extends AnyFunSuite with ScalatestRouteTest with F
   }
 
   test("return test results") {
-    saveProcessAndAssertSuccess(SampleProcess.process.id, SampleProcess.process)
-    val displayableProcess = ProcessConverter.toDisplayable(SampleProcess.process, TestProcessingTypes.Streaming, Category1)
     val testDataContent =
       """{"sourceId":"startProcess","record":"ala"}
         |{"sourceId":"startProcess","record":"bela"}""".stripMargin
-    val multiPart = MultipartUtils.prepareMultiParts("testData" -> testDataContent, "processJson" -> displayableProcess.asJson.noSpaces)()
-    Post(s"/processManagement/test/${SampleProcess.process.id}", multiPart) ~> withPermissions(deployRoute(), testPermissionDeploy |+| testPermissionRead) ~> check {
+    saveProcessAndAssertSuccess(SampleProcess.process.id, SampleProcess.process)
+
+    testScenario(SampleProcess.process, testDataContent) ~> check {
 
       status shouldEqual StatusCodes.OK
 
@@ -327,14 +326,10 @@ class ManagementResourcesSpec extends AnyFunSuite with ScalatestRouteTest with F
       .emptySink("end", "kafka-string", TopicParamName -> "'end.topic'", SinkValueParamName -> "''")
     val testDataContent =
       """{"sourceId":"startProcess","record":"ala"}
-        |{"sourceId":"startProcess","record":"bela"}""".stripMargin
-
+        |"bela"""".stripMargin
     saveProcessAndAssertSuccess(process.id, process)
 
-    val displayableProcess = ProcessConverter.toDisplayable(process, TestProcessingTypes.Streaming, Category1)
-
-    val multiPart = MultipartUtils.prepareMultiParts("testData" -> testDataContent, "processJson" -> displayableProcess.asJson.noSpaces)()
-    Post(s"/processManagement/test/${process.id}", multiPart) ~> withPermissions(deployRoute(), testPermissionDeploy |+| testPermissionRead) ~> check {
+    testScenario(process, testDataContent) ~> check {
       status shouldEqual StatusCodes.OK
     }
   }
@@ -350,16 +345,25 @@ class ManagementResourcesSpec extends AnyFunSuite with ScalatestRouteTest with F
           .source("startProcess", "csv-source")
           .emptySink("end", "kafka-string", TopicParamName -> "'end.topic'")
     }
-
     saveProcessAndAssertSuccess(process.id, process)
+    val tooLargeTestDataContentList = List((1 to 50).mkString("\n"), (1 to 50000).mkString("-"))
 
-    val displayableProcess = ProcessConverter.toDisplayable(process, TestProcessingTypes.Streaming, Category1)
-
-    List((1 to 50).mkString("\n"), (1 to 50000).mkString("-")).foreach { tooLargeData =>
-      val multiPart = MultipartUtils.prepareMultiParts("testData" -> tooLargeData, "processJson" -> displayableProcess.asJson.noSpaces)()
-      Post(s"/processManagement/test/${process.id}", multiPart) ~> withPermissions(deployRoute(), testPermissionDeploy |+| testPermissionRead) ~> check {
+    tooLargeTestDataContentList.foreach { tooLargeData =>
+      testScenario(process, tooLargeData) ~> check {
         status shouldEqual StatusCodes.BadRequest
       }
+    }
+  }
+
+  test("rejects test record with non-existing source") {
+    saveProcessAndAssertSuccess(SampleProcess.process.id, SampleProcess.process)
+    val testDataContent =
+      """{"sourceId":"startProcess","record":"ala"}
+        |{"sourceId":"unknown","record":"bela"}""".stripMargin
+
+    testScenario(SampleProcess.process, testDataContent) ~> check {
+      status shouldEqual StatusCodes.BadRequest
+      responseAs[String] shouldBe "Record 2 - scenario does not have source id: 'unknown'"
     }
   }